Rename org.mozilla.loginsapi => org.mozilla.sync15.logins
