+++ logins-api/android/LoginsApi/loginsapi/src/main/java/org/mozilla/sync15/logins/LoginsStorage.kt
@@ -6,7 +6,7 @@
  * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License. */
-package org.mozilla.loginsapi
+package org.mozilla.sync15.logins
 
 import java.io.Closeable
