+++ logins-api/android/LoginsApi/loginsapi/src/main/java/org/mozilla/sync15/logins/SyncResult.kt
@@ -6,7 +6,7 @@
  * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License. */
-package org.mozilla.loginsapi
+package org.mozilla.sync15.logins
 
 import java.util.ArrayList
 
@@ -111,7 +111,7 @@ class SyncResult<T> {
                 try {
                     result.completeFrom(valueListener.onValue(value))
                 } catch (ex: Exception) {
-                    result.completeFrom(SyncResult.fromException(ex))
+                    result.completeFrom(fromException(ex))
                 }
             }
