+++ logins-api/android/LoginsApi/loginsapi/src/test/java/org/mozilla/sync15/ExampleUnitTest.java
@@ -6,18 +6,20 @@
  * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License. */
-package org.mozilla.logins_example
+package org.mozilla.sync15;
 
-import android.support.v7.app.AppCompatActivity
-import android.os.Bundle
+import org.junit.Test;
 
-class StartupActivity : AppCompatActivity() {
+import static org.junit.Assert.*;
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_startup)
-
-        ExampleApp.instance.startNewIntent()
-        finish()
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
+ */
+public class ExampleUnitTest {
+    @Test
+    public void addition_isCorrect() {
+        assertEquals(4, 2 + 2);
     }
 }
