+++ logins-api/android/LoginsSampleApp/app/src/main/java/org/mozilla/sync15_logins_example/LoginActivity.kt
@@ -6,7 +6,7 @@
  * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License. */
-package org.mozilla.logins_example
+package org.mozilla.sync15_logins_example
 
 import android.net.Uri
 import android.support.v7.app.AppCompatActivity
