+++ logins-api/android/LoginsSampleApp/app/src/main/java/org/mozilla/sync15_logins_example/StartupActivity.kt
@@ -6,10 +6,18 @@
  * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License. */
-package org.mozilla.loginsapi
+package org.mozilla.sync15_logins_example
 
-// TODO: Get more descriptive errors here.
-class LoginsStorageException(msg: String): Exception(msg)
+import android.support.v7.app.AppCompatActivity
+import android.os.Bundle
 
-// This doesn't really belong in this file...
-class MismatchedLockException(msg: String): Exception(msg)
+class StartupActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_startup)
+
+        ExampleApp.instance.startNewIntent()
+        finish()
+    }
+}
