+++ logins-api/android/LoginsSampleApp/app/src/test/java/org/mozilla/sync15_logins_example/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package org.mozilla.logins_example
+package org.mozilla.sync15_logins_example
 
 import org.junit.Test
